Fixes related with PR review
