+++ tests/framework/console/FakeHelpController.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace yiiunit\framework\console;
+
+use yii\console\controllers\HelpController;
+
+class FakeHelpController extends HelpController
+{
+    private static $_actionIndexLastCallParams;
+
+    public function actionIndex($command = null)
+    {
+        self::$_actionIndexLastCallParams = func_get_args();
+    }
+
+    public static function getActionIndexLastCallParams()
+    {
+        $params = self::$_actionIndexLastCallParams;
+        self::$_actionIndexLastCallParams = null;
+
+        return $params;
+    }
+}
